ephemeris/tool/pyproject.toml
[project]
name = "ephemerid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astropy>=7.1.1",
    "jplephem>=2.23",
    "pillow>=12.0.0",
]

[tool.pyright]
venv = ".venv"
venvPath = "."
executionEnvironments = [{"root" = "."}]
